feat(common, ..): add system param to enforce uniqueness
