+++ src/plugins/common/src/lib.rs
@@ -7,6 +7,7 @@ pub mod errors;
 pub mod observers;
 pub mod resources;
 pub mod states;
+pub mod system_params;
 pub mod systems;
 pub mod tools;
 pub mod traits;
