+++ src/plugins/physics/src/system_params/ray_caster/set_world_camera.rs
@@ -12,11 +12,11 @@ impl<'w, 's> NoWorldCameraSet for RayCasterMut<'w, 's> {
 	type TSetter = WorldCameraSetter<'w, 's>;
 
 	fn no_world_camera_set(self) -> Option<Self::TSetter> {
-		if !self.inner.world_cameras.is_empty() {
+		if !self.world_cameras.is_empty() {
 			return None;
 		}
 
-		Some(WorldCameraSetter(self.inner.commands))
+		Some(WorldCameraSetter(self.commands))
 	}
 }
